requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonfixer==0.2.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21
langdetect==1.0.9
langsmith==0.0.87
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.5.0
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
PyMuPDF==1.23.20
PyMuPDFb==1.23.9
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
replicate==0.22.0
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.1
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
